feat: migrate contracts to it's own package
