+++ .changeset/great-jokes-return.md
@@ -1,0 +1,5 @@
+---
+'@sovryn/ethers-provider': patch
+---
+
+SOV-1016: migrate network map to the ethers-provider package
