+++ .changeset/heavy-kangaroos-prove.md
@@ -1,0 +1,5 @@
+---
+'@sovryn/contracts': patch
+---
+
+SOV-1016: migrate contracts with abis to the package
