+++ .changeset/wicked-pets-peel.md
@@ -1,0 +1,5 @@
+---
+'frontend': patch
+---
+
+SOV-1016: change how chains are defined and used
