+++ apps/frontend/package.json
@@ -3,7 +3,8 @@
   "version": "0.1.4",
   "private": true,
   "dependencies": {
-    "@sovryn/ethers-provider": "*",
+    "@sovryn/ethers-provider": "0.0.0",
+    "@sovryn/contracts": "0.0.0",
     "@sovryn/onboard-common": "0.0.1",
     "@sovryn/onboard-core": "0.0.2",
     "@sovryn/onboard-injected": "0.0.1",
