+++ apps/frontend/src/config/chains.ts
@@ -7,26 +7,29 @@ export const chains: Chain[] = [
   {
     id: ChainIds.RSK_MAINNET,
     label: 'Rootstock',
-    symbol: 'RBTC',
+    token: 'RBTC',
     rpcUrl: ['https://mainnet.sovryn.app/rpc', 'https://public-node.rsk.co'],
+    blockExplorerUrl: 'https://explorer.rsk.co',
   },
   {
     id: ChainIds.MAINNET,
     label: 'Ethereum',
-    symbol: 'ETH',
+    token: 'ETH',
     rpcUrl: [
       'https://main-light.eth.linkpool.io',
       'https://cloudflare-eth.com/',
       'https://api.mycryptoapi.com/eth',
     ],
+    blockExplorerUrl: 'https://etherscan.io',
   },
   {
     id: ChainIds.RSK_TESTNET,
     label: 'Rootstock testnet',
-    symbol: 'tRBTC',
+    token: 'tRBTC',
     rpcUrl: [
       'https://testnet.sovryn.app/rpc',
       'https://public-node.testnet.rsk.co',
     ],
+    blockExplorerUrl: 'https://explorer.testnet.rsk.co',
   },
 ];
