+++ packages/contracts/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "@sovryn/contracts",
+  "version": "0.0.0",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "license": "MIT",
+  "author": {
+    "name": "Victor Creed",
+    "email": "victor@sovryn.app"
+  },
+  "scripts": {
+    "dev": "tsup --watch",
+    "build": "tsup",
+    "test": "jest --passWithNoTests",
+    "test:staged": "jest --passWithNoTests --bail --onlyChanged",
+    "coverage": "jest --coverage",
+    "lint": "eslint ./src/**/*.ts"
+  },
+  "dependencies": {
+    "@sovryn/ethers-provider": "0.0.0",
+    "ethers": "5.7.1",
+    "lodash.get": "4.4.2",
+    "lodash.set": "4.3.2"
+  },
+  "devDependencies": {
+    "@babel/preset-typescript": "7.18.6",
+    "jest": "29.0.3",
+    "tsup": "6.2.3",
+    "typescript": "4.8.3"
+  },
+  "peerDependencies": {
+    "ethers": "*"
+  },
+  "files": [
+    "dist"
+  ]
+}
