+++ packages/contracts/tsup.config.js
@@ -1,0 +1,10 @@
+import { defineConfig } from 'tsup';
+
+export default defineConfig({
+  entry: ['src/index.ts'],
+  splitting: true,
+  sourcemap: true,
+  clean: true,
+  dts: true,
+  treeshake: true,
+});
